% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{data_sieve}
\alias{data_sieve}
\title{Cleanse dataset of low_effect samples}
\usage{
data_sieve(
  x,
  y,
  coeffs = c(0.7, 0.9, 750),
  gap_percentage = 0.01,
  if_plot = TRUE
)
}
\arguments{
\item{x}{data x of independent variable}

\item{y}{data y of effects to be sieved}

\item{coeffs}{a vector of length 3: compared quantile, limited percentage error, limited direct error}

\item{gap_percentage}{number of neighbors to be selected}

\item{if_plot}{whether or not to draw a plot}
}
\value{
cleansed dataset
}
\description{
Cleanse dataset of low_effect samples
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
temp1 = data_sieve(temp[,1],temp[,2],c(0.7,0.85,500))
}
