% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{disperse_level}
\alias{disperse_level}
\title{For a 2-dimension plot, compare the level how the plot dispersed}
\usage{
disperse_level(x_data, y_data, nx = 5, ny = 5)
}
\arguments{
\item{x_data}{data of dimension 1}

\item{y_data}{data of dimension 2}

\item{nx}{how many blocks in x-axis}

\item{ny}{how many blocks in y-axis}
}
\value{
a real number corresponding how much it disperses
}
\description{
For a 2-dimension plot, compare the level how the plot dispersed
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
disperse_level(temp[,1],temp[,2]) #3.71
temp = data_generator(1000,55,gen_type = 'normal')
disperse_level(temp[,1],temp[,2]) #3.10
}
