% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{draw_lines_model}
\alias{draw_lines_model}
\title{Draw main predict lines in EXISTING plot with model generated by function 'SLM_construct'}
\usage{
draw_lines_model(one_model, model_col = "blue", x_beg = -100, x_end = 100)
}
\arguments{
\item{one_model}{the model generated by function 'SLM_construct'}

\item{model_col}{color of line for the drawing}

\item{x_beg}{lowest x value for drawing, default -100}

\item{x_end}{largest x value for drawing, default 100}
}
\value{
0
}
\description{
Draw main predict lines in EXISTING plot with model generated by function 'SLM_construct'
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
model_1 = SLM_construct(temp[,1],temp[,2],fixed_psi = c(2,6,8,10))
dev.off()
draw_lines_model(model_1,'red')
plot(temp[,1],temp[,2])
draw_lines_model(model_1,'red')
}
