% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{L1_shrink}
\alias{L1_shrink}
\title{Loss function of L(x) = I(abs(x)>m)*(abs(x)-m)}
\usage{
L1_shrink(x, m)
}
\arguments{
\item{x}{original error}

\item{m}{the position of gap}
}
\value{
loss
}
\description{
Loss function of L(x) = I(abs(x)>m)*(abs(x)-m)
}
\examples{
plot(1:100,L1_shrink(1:100,30))
}
