% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{L2}
\alias{L2}
\title{Loss function of L(x) = x^2}
\usage{
L2(x)
}
\arguments{
\item{x}{original error}
}
\value{
loss
}
\description{
Loss function of L(x) = x^2
}
\examples{
plot(1:100,L2(1:100))
}
