% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{msl_network}
\alias{msl_network}
\title{Generate a model of multivariate segmented linear network}
\usage{
msl_network(
  x,
  y,
  cuts,
  init_deltas = NULL,
  iter_ctrl = NULL,
  loss = L2,
  params = NULL,
  lower = NULL
)
}
\arguments{
\item{x}{independent variables}

\item{y}{dependent variable}

\item{cuts}{list, modeling with segmented line by points of ({cuts[i,j,k...]}, deltas[i])}

\item{init_deltas}{default to be mean(y)}

\item{iter_ctrl}{as 'control' in function 'nlminb'}

\item{loss}{loss function}

\item{params}{loss function}

\item{lower}{lower bound of deltas, default -Inf for values and 0.5 for metric}
}
\value{
a model of multivariate segmented linear network
}
\description{
Generate a model of multivariate segmented linear network
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
y = temp[,2]
one_trial = temp[,-2]
cuts = list(c(-2,4,8,12,21),c(-10,3,16,30),c(0,33,67,100))
output = msl_network(x = one_trial, y = y, cuts = cuts,
    loss = L1, iter_ctrl = list(iter.max=4, trace = 1))
}
