% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi_sum_loss}
\alias{multi_sum_loss}
\title{Calculate aggregated loss with input loss function}
\usage{
multi_sum_loss(deltas, x, y, cuts, loss, params = NULL)
}
\arguments{
\item{deltas}{modeling with segmented line by points of ({cuts[i,j,k...]}, deltas[i])}

\item{x}{independent variable}

\item{y}{dependent variable}

\item{cuts}{list, modeling with segmented line by points of ({cuts[i,j,k...]}, deltas[i])}

\item{loss}{loss function}

\item{params}{Other parameters in loss function. Use 'NULL' if only one input.}
}
\value{
total loss
}
\description{
Calculate aggregated loss with input loss function
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
y = temp[,2]
one_trial = temp[,-2]
cuts = list(c(-2,3,5,8,11,15,30),c(-10,0,10,20,30),c(0,33,67,100))
deltas = c(rep(500,7*5*4),0.1,0.1)
output = nlminb(deltas, multi_sum_loss, x = one_trial, y = y, cuts = cuts,
    loss = L2, control = list(iter.max=4, eval.max = 45,
    trace = 1, abs.tol = 1e-20), lower = c(rep(-Inf,140),0,0))
multi_sum_loss(output$par,one_trial, y, cuts, L2)
}
