% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{predict_model}
\alias{predict_model}
\title{Predict with SLM model generated by function 'SLM_construct'}
\usage{
predict_model(one_model, newdata)
}
\arguments{
\item{one_model}{SLM model generated by function 'SLM_construct'}

\item{newdata}{new independent variable to be predicted}
}
\value{
predicted values
}
\description{
Predict with SLM model generated by function 'SLM_construct'
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
model_1 = SLM_construct(temp[,1],temp[,2],fixed_psi = c(2,6,8,10))
predict_model(model_1, c(0,1,10,100,0.5,-1333))
model_2 = SLM_construct(temp[,c(1,3,4)],temp[,2],fixed_psi = c(2,6,8,10))
predict_model(model_2, newdata = temp[1:5,])
}
