% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{sim_sigmoid_hard}
\alias{sim_sigmoid_hard}
\title{Loss function of L(x) = I(abs(x)>m), with gap's tangent -> inf}
\usage{
sim_sigmoid_hard(x, m)
}
\arguments{
\item{x}{original error}

\item{m}{the position of gap}
}
\value{
loss
}
\description{
Loss function of L(x) = I(abs(x)>m), with gap's tangent -> inf
}
\examples{
plot(1:100,sim_sigmoid_hard(1:100,30))
}
