% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{sim_sigmoid_soft}
\alias{sim_sigmoid_soft}
\title{Loss function of L(x) = (1+exp(-k*((x/m)^2)-1))^-1}
\usage{
sim_sigmoid_soft(x, k = c(3, 100))
}
\arguments{
\item{x}{original error}

\item{k}{a length 2 vector, indicates how sharp the gap is & the position of gap}
}
\value{
loss
}
\description{
Loss function of L(x) = (1+exp(-k*((x/m)^2)-1))^-1
}
\examples{
plot(1:10,sim_sigmoid_soft(1:10,c(3,5)))
plot(1:100,sim_sigmoid_soft(1:100,c(3,50)))
plot(1:100,sim_sigmoid_soft(1:100,c(10,50)))
}
