% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{SLM_construct}
\alias{SLM_construct}
\title{Segmented linear model main construct function (one variable)}
\usage{
SLM_construct(
  x,
  y,
  fixed_psi,
  loss_function = L2,
  params = NULL,
  init_val = NULL
)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{fixed_psi}{segment location(cuts)}

\item{loss_function}{What loss function to use. L2 in default}

\item{params}{The parameters in loss function. For L2 loss, NULL in default}

\item{init_val}{Vector of length(fixed_psi)+2. Default: c(rep(mean(y),length(fixed_psi)),0,0). For the corresponding y_hat at each fixed_psi, and initial & final tangent value}
}
\value{
a list with segmented location, corresponding y_hat, tangents and total loss
}
\description{
Segmented linear model main construct function (one variable)
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
SLM_construct(temp[,1],temp[,2],fixed_psi = c(2,6,8,10))
}
