% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{sum_loss}
\alias{sum_loss}
\title{Calculate aggregated loss with input loss function}
\usage{
sum_loss(deltas, x, y, cuts, loss, params = NULL)
}
\arguments{
\item{deltas}{modeling with segmented line by points of (cuts[i], deltas[i]), plus initial and final tangents}

\item{x}{independent variable}

\item{y}{dependent variable}

\item{cuts}{modeling with segmented line by points of (cuts[i], deltas[i])}

\item{loss}{loss function}

\item{params}{Other parameters in loss function. Use 'NULL' if only one input.}
}
\value{
Summation of losses for all samples
}
\description{
Calculate aggregated loss with input loss function
}
\examples{
temp = data_generator(1000,55,gen_type = 'chaos&cut')
sum_loss(c(10,200,350,0,0),temp[,1],temp[,2],c(5,10,15),sim_sigmoid_soft, c(3,100))
nlminb(c(10,200,350,0,0),sum_loss, x=temp[,1],y=temp[,2], cuts=c(5,10,15),params=c(5,100),loss=sim_sigmoid_soft)
}
